warnings corregidos

+++ Flop/Flop/Enemy1.cpp
@@ -2,7 +2,7 @@
 
 Enemy1::Enemy1() {
 	_rec.x = 880.0f;
-	_rec.y = GetRandomValue(40, 400);
+	_rec.y = static_cast<float>(GetRandomValue(40, 400));
 	_rec.height = 40.0f;
 	_rec.width = 40.0f;
 	_color = RED;
@@ -18,7 +18,7 @@ void Enemy1::update() {
 	_rec.x -= 5.0f;
 	if (_rec.x <= -40.0f) {
 		_rec.x = 880.0f;
-		_rec.y = GetRandomValue(40, 400);
+		_rec.y = static_cast<float>(GetRandomValue(40, 400));
 	}
 }
 void Enemy1::GroundUpdate() {
